feat: Add formatted size labels and load builds based on active operating system

+++ MarmaladeLauncher/Models/InstallationEntry.cs
@@ -1,8 +1,10 @@
+using MarmaladeLauncher.Utils;
+
 namespace MarmaladeLauncher.Models;
 
 public class InstallationEntry {
     public string id { get; set; } = string.Empty;
     public string url { get; set; } = string.Empty; // download url
     public string date { get; set; } = string.Empty;
-    public int size { get; set; } = 0;
+    public long size { get; set; } = 0;
 }
